+++ R/shinyML_regression.R
@@ -33,8 +33,9 @@
 #' @importFrom plotly plotlyOutput renderPlotly ggplotly plot_ly layout add_trace
 #' @importFrom shinyWidgets materialSwitch switchInput sendSweetAlert knobInput awesomeCheckbox actionBttn prettyCheckboxGroup
 #' @importFrom shinyjs useShinyjs hideElement
-#' @importFrom stats predict reorder cor
+#' @importFrom stats predict reorder cor acf
 #' @importFrom lubridate is.Date is.POSIXct
+#' @importFrom graphics par
 #' @author Jean Bertin, \email{jean.bertin@mines-paris.org}
 #' @export
 
@@ -213,15 +214,16 @@ shinyML_regression <- function(data = data,y,framework = "h2o", share_app = FALS
                                         argonTab(tabName = "Variables Summary",
                                                  active = FALSE,
                                                  fluidRow( 
-                                                   argonColumn(width = 6,
+                                                   argonColumn(width = 4,
                                                                br(),
                                                                br(),
                                                                withSpinner(DTOutput("variables_class_input", height = "100%", width = "100%"))
                                                    ),
-                                                   argonColumn(width = 6,
+                                                   argonColumn(width = 8,
                                                                div(align = "center",
-                                                                   radioButtons(inputId = "input_var_graph_type",label = "",choices = c("Boxplot","Histogram"),selected = "Boxplot",inline = T)
+                                                                   radioButtons(inputId = "input_var_graph_type",label = "",choices = c("Boxplot","Histogram","Autocorrelation"),selected = "Boxplot",inline = T)
                                                                ),
+                                                               div(align = "center",uiOutput("message_autocorrelation")),
                                                                withSpinner(plotlyOutput("variable_boxplot", height = "100%", width = "100%")))
                                                  )
                                                  
@@ -750,6 +752,15 @@ shinyML_regression <- function(data = data,y,framework = "h2o", share_app = FALS
       
     })
     
+    # Define indicating that autocorrelation plot is only available for time series
+    output$message_autocorrelation <- renderUI({
+      
+      points_serie <-eval(parse(text = paste0("data[,",colnames(data)[input$variables_class_input_rows_selected],"]"))) 
+      if (input$input_var_graph_type == "Autocorrelation" & !is.numeric(points_serie)){
+        argonH1("Only available for numerical variables",display = 4)
+      }
+    })
+
     # Make glm parameters correspond to cursors and radiobuttons choices when user click on "Run generalized linear regression" button 
     observeEvent(input$run_glm,{
       
@@ -913,17 +924,26 @@ shinyML_regression <- function(data = data,y,framework = "h2o", share_app = FALS
     # Define boxplot corresponding to  selected variable in variables_class_input 
     output$variable_boxplot <- renderPlotly({
       
+      par("mar")
+      par(mar=c(1,1,1,1))
+      
       column_name <- colnames(data)[input$variables_class_input_rows_selected]
-      
-      if (input$input_var_graph_type == "Histogram"){chart_type <- "histogram"}
-      else if (input$input_var_graph_type == "Boxplot"){chart_type <- "box"}
-      
-      plot_ly(x = eval(parse(text = paste0("data[,",column_name,"]"))),
-              type = chart_type,
-              name = column_name
-      )
-      
-      
+      points_serie <-eval(parse(text = paste0("data[,",column_name,"]"))) 
+        
+      if (input$input_var_graph_type == "Histogram"){
+        plot_ly(x = points_serie,type = "histogram",name = column_name)
+        }
+      else if (input$input_var_graph_type == "Boxplot"){
+        plot_ly(x = points_serie,type = "box",name = column_name)
+        }
+      else if (input$input_var_graph_type == "Autocorrelation"){
+        req(is.numeric(points_serie))
+        acf_object <- acf(points_serie,lag.max = 100)
+        data_acf <- cbind(acf_object$lag,acf_object$acf) %>% as.data.table() %>% setnames(c("Lag","ACF"))
+        plot_ly(x = data_acf$Lag, y = data_acf$ACF, type = "bar")
+
+        }
+
     })
     
     # Define plotly chart to explore dependencies between variables 
@@ -934,7 +954,7 @@ shinyML_regression <- function(data = data,y,framework = "h2o", share_app = FALS
       req(!is.null(input$y_variable_input_curve))
       
       if (input$checkbox_time_series == TRUE){
-        data_train_chart <- eval(parse(text = paste0("data[",input$time_serie_select_column," >= input$train_selector[2],][",input$time_serie_select_column," <= input$train_selector[2],]")))
+        data_train_chart <- eval(parse(text = paste0("data[",input$time_serie_select_column," >= input$train_selector[1],][",input$time_serie_select_column," <= input$train_selector[2],]")))
         data_test_chart <- eval(parse(text = paste0("data[",input$time_serie_select_column," > input$test_selector[1],][",input$time_serie_select_column," <= input$test_selector[2],]")))
         
       }